Factorize common code
